#84 Enhance Measurements module pages
